fixed warnings

+++ source/utilities.cpp
@@ -5,7 +5,7 @@
 int readSizeOfFile( FILE * file )
 {
     fseek(file, 0L, SEEK_END);
-    int sizeOfFile = ftell(file) + 1;
+    int sizeOfFile = (int)ftell(file) + 1;
 
     fseek(file, 0L, SEEK_SET);
 
@@ -16,7 +16,7 @@ void printText( char ** data, size_t numOfLines, FILE * outputFile )
 {
     CASSERT(outputFile);
 
-    for (int numOfReadedLines = 0; numOfReadedLines < numOfLines; numOfReadedLines++) {
+    for (size_t numOfReadedLines = 0; numOfReadedLines < numOfLines; numOfReadedLines++) {
         fputs(data[numOfReadedLines], outputFile);
         fputc('\n', outputFile);
     }
